requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==3.2.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.111.0
fastapi-cli==0.0.23
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.2.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.2.0
pytest-asyncio==0.23.7
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==5.0.4
rich==14.3.3
rich-toolkit==0.19.4
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tomli==2.4.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.1
